Refactoring for thread-safety and reduction of complexity
The original behaviour is retained, but the implementation, apart from 
keeping the state pattern, has the following significant changes:

* Stop using atomics as the combined use of closely related atomics is 
not thread-safe. It led to non-deterministic behaviour, in particular a 
state could execute its logic without being the current state anymore.
  
* Reduce mutability, from 5 mutable fields to 2, through remodeling the 
states.

* Introduce a 4th state (`UnstableState`), to avoid having to manually 
deal with `turbulenceStartedAt` timings etc.
  
* Rename `UnhealthyState` to `UnavailableState` to avoid confusing it 
with the new `UnstableState`. 
  
* Move all state and behaviour only needed in one of the states to the 
corresponding state implementation.

* Introduce `CountingState` and `TimedState` traits to avoid duplicate 
logic.
  
* Move separate `Executor` and states into `CircuitBreaker` 
implementation as its reduction in size made the separation obsolete 
(which also removes the indirection to talk through a 
`CircuitBreakerModel`).
  
* Remove the Registry. It was not thread-safe either and there does not 
seem to be any benefit to organizing `CircuitBreaker`s in a registry.
  
* `UnavailableState` does no longer throw an Exception in its 
`processResult` method. Due to the asynchronous nature of the circuit 
breaker, this can happen any time.

* Introduce `CircuitBreakerConfig` case class to avoid the need to pass 
all configuration values around individually.
  
* Introduce a custom predicate (`Throwable => Boolean`) to tell the 
`CircuitBreaker` which exceptions are expected results (e.g. a 404 for a 
lookup) and should not be treated as a failure.